app/requirements.txt
fastapi==0.112.0
openai==1.65.5
pydantic==2.8.2
pymongo==4.8.0
python-dotenv==1.0.1
uvicorn==0.30.5
gunicorn==23.0.0
slowapi==0.1.9
PyJWT==2.9.0
cryptography==43.0.0
sympy==1.13.3
mcp==1.1.2